refactor(desktop): extract start session strategies
